master dotfiles folder with env based sourcing

+++ bootstrap.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-for file in {exports,functions,aliases,completion,prompt,extra,cygwin}; do
+for file in {bash_profile,bashrc,dotenv,inputrc,screenrc,wgetrc}; do
 	[ -f "$HOME/.$file" ] && mv "$HOME/.$file" "$HOME/.$file.bak"
-	ln -s "$PWD/.$file" "$HOME/.$file"
+	[ -f "$PWD/.$file" ] && ln -s "$PWD/.$file" "$HOME/.$file"
 done
-unset file
